[refactor/#10] Refactor package and class names, update imports accordingly
