+++ src/main/java/com/api/meetudy/study/group/service/GroupRetrievalService.java
@@ -5,8 +5,8 @@ import com.api.meetudy.study.group.dto.StudyGroupDto;
 import com.api.meetudy.study.group.entity.StudyGroup;
 import com.api.meetudy.study.group.entity.StudyGroupMember;
 import com.api.meetudy.study.group.mapper.StudyGroupMapper;
-import com.api.meetudy.study.group.repository.StudyGroupMemberRepository;
-import com.api.meetudy.study.group.repository.StudyGroupRepository;
+import com.api.meetudy.study.group.repository.GroupMemberRepository;
+import com.api.meetudy.study.group.repository.GroupRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -17,8 +17,8 @@ import java.util.List;
 @RequiredArgsConstructor
 public class GroupRetrievalService {
 
-    private final StudyGroupRepository groupRepository;
-    private final StudyGroupMemberRepository groupMemberRepository;
+    private final GroupRepository groupRepository;
+    private final GroupMemberRepository groupMemberRepository;
     private final StudyGroupMapper studyGroupMapper;
 
     @Transactional(readOnly = true)
