+++ src/main/java/com/api/meetudy/study/recommendation/service/RecommendationService.java
@@ -5,8 +5,8 @@ import com.api.meetudy.study.group.dto.StudyGroupDto;
 import com.api.meetudy.study.group.entity.StudyGroup;
 import com.api.meetudy.study.group.entity.StudyGroupMember;
 import com.api.meetudy.study.group.mapper.StudyGroupMapper;
-import com.api.meetudy.study.group.repository.StudyGroupMemberRepository;
-import com.api.meetudy.study.group.repository.StudyGroupRepository;
+import com.api.meetudy.study.group.repository.GroupMemberRepository;
+import com.api.meetudy.study.group.repository.GroupRepository;
 import com.api.meetudy.study.recommendation.entity.Interest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -18,8 +18,8 @@ import java.util.stream.Collectors;
 @RequiredArgsConstructor
 public class RecommendationService {
 
-    private final StudyGroupRepository groupRepository;
-    private final StudyGroupMemberRepository groupMemberRepository;
+    private final GroupRepository groupRepository;
+    private final GroupMemberRepository groupMemberRepository;
     private final StudyGroupMapper studyGroupMapper;
 
     public List<StudyGroupDto> recommendStudyGroups(Member member) {
